[#188]refactor:네비게이션바 공통컴포넌트로 이동
