+++ src/components/Common/NavigationBar/Styles.js
