+++ src/pages/Booth/Booth.tsx
@@ -3,7 +3,7 @@ import * as S from './Styles';
 import TypeButton from '../../components/Booth/TypeButton/TypeButton';
 import Booths from '../../components/Booth/Booths/Booths';
 import Header from '../../components/Booth/Header/Header';
-import NavigationBar from 'components/NavigationBar/NavigationBar';
+import NavigationBar from 'components/Common/NavigationBar/NavigationBar';
 import GotoMeeting from 'components/GotoMeeting/GotoMeeting';
 
 const Booth: React.FC = () => {
