+++ src/pages/Booth/BoothDetail/BoothDetail.tsx
@@ -6,7 +6,7 @@ import BackHeader from '../../../components/BoothDetail/BackHeader/BackHeader';
 import AccountCopy from '../../../components/BoothDetail/AccountCopy/AccountCopy';
 import ItemInventory from 'components/BoothDetail/ItemInventory/ItemInventory';
 import Explanation from 'components/BoothDetail/Explanation/Explanation';
-import NavigationBar from 'components/NavigationBar/NavigationBar';
+import NavigationBar from 'components/Common/NavigationBar/NavigationBar';
 import GotoMeeting from 'components/GotoMeeting/GotoMeeting';
 import { GetboothDetail } from '../../../api/booth/GetboothDetail';
 import { boothDetailState } from '../../../recoil/state/boothState';
