+++ src/pages/OurTeamIntro/OurTeamIntro.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import * as S from './Styles';
 import Header from 'components/OurTeamIntro/Header/Header';
 import MyProfile from 'components/Common/Profile/MyProfile';
-import NavigationBar from 'components/NavigationBar/NavigationBar';
+import NavigationBar from 'components/Common/NavigationBar/NavigationBar';
 import TipModal from 'components/OurTeamIntro/Modal/TipModal/TipModal';
 import { getOurTeamDetail } from '../../api/Meeting/GetourTeamDetail';
 import { useLocation } from 'react-router-dom';
